requirements.txt
Django==6.0.4
django-environ==0.13.0
django-q2==1.10.0
PlexAPI==4.18.1
requests==2.33.1
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.2.2
django-picklefield==3.4.0
tzdata==2026.2
redis==5.0.1
